+++ reader.h
@@ -10,7 +10,7 @@
 #include <fstream>
 
 std::vector<double> parse_ica(std::string filename);
-std::vector<long double> parse_jst(std::string filename);
+std::vector<double> parse_jst(std::string filename);
 std::vector<std::string> parse_new_plates(std::string filename);
 
 #endif
